% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleNoBounds-package.r
\docType{package}
\name{nimbleNoBounds}
\alias{nimbleNoBounds}
\title{nimbleNoBounds}
\description{
A collection of common bounded univariate probability distributions transformed to the real line for unbounded MCMC sampling.
}
\section{Transformed probability distributions for unbounded sampling in nimble.}{
NA
}

\seealso{
\url{https://r-nimble.org/}
}
