% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleNoBounds.R
\docType{package}
\name{nimbleNoBoundss}
\alias{nimbleNoBoundss}
\title{nimbleNoBoundss}
\description{
A collection of nimble functions for sampling common probability distributions on unbounded scales.
}
\section{Transformed probability distributions for unbounded sampling in nimble.}{
NA
}

\seealso{
\url{https://r-nimble.org/}
}
