% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLogChisq.R
\name{dLogChisq}
\alias{dLogChisq}
\alias{rLogChisq}
\title{Log transformed chi-squared distribution.}
\usage{
dLogChisq(x, df = 1, log = 0)

rLogChisq(n = 1, df = 1)
}
\arguments{
\item{x}{A continuous random varaible on the real line, where y=exp(x) and y ~ dchsq(df).}

\item{df}{df parameter of y ~ dchisq(df).}

\item{log}{Logical flag to toggle returning the log density.}

\item{n}{Number of random variables. Currently limited to 1, as is common in nimble. See ?replicate for an alternative.}
}
\value{
The density or log density of x, such that x = log(y) and y ~ dchisq(df).
}
\description{
\code{dLogChisq} and \code{rLogChisq} provide a log-transformed chi-squared distribution.
}
\examples{

n  = 100000
df = 3
y  = rchisq(n=n, df=df)
x  = log(y)

par(mfrow=n2mfrow(2))
## Plot 1
hist(x, n=100, freq=FALSE)
curve(dLogChisq(x, df=df), -15, 4, n=1001, col="red", add=TRUE, lwd=3)
## Plot 2: back-transformed
xNew = replicate(n=n, rLogChisq(n=1, df=df))
yNew = exp(xNew)
hist(yNew, n=100, freq=FALSE, xlab="exp(x)")
curve(dchisq(x, df=df), 0, 4, n=1001, col="red", lwd=3, add=TRUE)

code = nimbleCode({
  x ~ dLogChisq(df = 0.5)
})
modelR = nimbleModel(code=code)
modelC = compileNimble(modelR)
conf  = configureMCMC(modelC)
mcmc  = buildMCMC(conf=conf)
cMcmc = compileNimble(mcmc)
x = as.vector(runMCMC(mcmc=cMcmc, niter=50000))
par(mfrow=n2mfrow(3))
## Plot 1: MCMC trajectory
plot(x, typ="l")
## Plot 2: taget density on unbounded sampling scale
hist(x, n=100, freq=FALSE)
curve(dLogChisq(x, df=0.5), -15, 5, n=1001, col="red", lwd=3, add=TRUE)
## Plot 3: taget density on bounded scale
hist(exp(x), n=100, freq=FALSE)
curve(dchisq(x, df=0.5), 0, 25, n=1001, col="red", lwd=3, add=TRUE)
}
\author{
David R.J. Pleydell
}
