% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleNoBounds.R
\name{nimbleNoBounds}
\alias{nimbleNoBounds}
\title{nimbleNoBounds}
\description{
A collection of NIMBLE functions for sampling common probability distributions on unbounded scales.
}
\section{Transformed probability distributions for unbounded sampling in NIMBLE.}{
NA
}

\seealso{
\url{https://r-nimble.org/}
_PACKAGE
}
